requirements.txt
appdirs==1.4.3
boto3==1.4.4
botocore==1.5.45
Django==1.11
django-bakery==0.10.2
django-storages==1.5.2
docutils==0.13.1
futures==3.1.1
jmespath==0.9.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
s3transfer==0.1.10
six==1.10.0
